sales_engineering/POC/VideoLogoDetection/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
av
clarifai
clarifai-grpc
clarifai-protocol
cryptography==41.0.7
ffmpeg-python==0.2.0
google-generativeai==0.3.1
loguru==0.7.2
marshmallow==3.20.1
moviepy==2.1.2
numpy==1.26.4
opencv-python-headless==4.8.1.78
passlib==1.7.4
Pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.5.2
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.20
requests==2.32.3
rq==1.15.1
tqdm==4.66.1
Werkzeug==3.1.3
pytest==7.4.4